make sure environment is set to 'development'

+++ seeder/teardown.ts
@@ -2,6 +2,9 @@ import { prisma } from '.'
 
 
 export const clearUsers = async () => {
+    if(process.env.ENVIRONMENT !== 'development'){
+        return
+    }
     const response = await prisma.user.findMany()
     if(response.length > 0){
     await prisma.refreshToken.deleteMany()
